+++ lib/src/screens/image/utils/positions_utils.dart
@@ -38,7 +38,7 @@ class PositionsUtils {
       var pos = arr[i];
       var radiusRatio = face.radius / pos.getVisibleRadius();
       if (radiusRatio > 0.85 && radiusRatio < 1.15) {
-        if (pos.isInnerPoint(face.x, face.y)) {
+        if (pos.isInnerPoint(face.x.toDouble(), face.y.toDouble())) {
           var diffX = (face.x - pos.posX).abs();
           var diffY = (face.y - pos.posY).abs();
           if ((diffX + diffY) < (face.radius * 0.5)) {
